fix: use of size_t, not long but with cast when using erase.

+++ src/hash_set_refinable.h
@@ -53,9 +53,9 @@ class HashSetRefinable : public HashSetBase<T> {
 
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (long i = 0; i < bucket.size(); i++) {
+    for (size_t i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
-        bucket.erase(bucket.begin() + i);
+        bucket.erase(bucket.begin() + static_cast<long>(i));
         return true;
       }
     }
